Fixed possible inclusion of empty Accept header in SUBSCRIBE.
Fixed correct proxying of REGISTER/SUBSCRIBE etc where have an AOR domain which resolves to same IP address as an explicit host name specified. That is aor=sip:fred@bloggs.com and remote=sip.bloggs.com should mean use the host sip.bloggs.com and do not do DNS SRV record checks etc. But it is common for bloggs.com and sip.bloggs.com to resolve to the same IP address and that was what was beoing matched in OPAL. Not it does a simple caseless text compare.

git-svn-id: http://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_10@25406 023b2edf-31b2-4de3-b41e-bca80c47788f