Merged revision(s) 26906 from opal/trunk: Repaired regression damage caused by revision 26843 which "Fixed correct interpretation of ';' parameters on a SIP header field URI (e.g. To or From) where no '<'/'>' are present. RFC EBNF indicates parameters then belong to the field not the URI."
The problem is a 1000 other places use the old semantics, so had to make all the default parsing do it that way or things like sip:addr;transport=tcp are misinterpreted. Now, a special parser is used when the source string is a MIME field, so all of the existing assignments work as the used to. Have to remember to do that in the future.
........


git-svn-id: http://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_10@26908 023b2edf-31b2-4de3-b41e-bca80c47788f