Fix for bug ID: 3495610 "Incorrect handling of SRV records for PUBLISH", reported by Anonymous
RFC 5509 indicates that the return value from _pres._sip.foo.org is in fact another domain, which should then be processed via the _sip._udp.foo.org SRV record. That is, two look ups are required.

Also if the first fails, we should then try the second with the original domain.

git-svn-id: http://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_10@27081 023b2edf-31b2-4de3-b41e-bca80c47788f