Fixed loss of media if do multiple hold/retrieve on H.323. Some old OpenH323 hold handling for Music on Hold getting in the way.
Added pause of transmit media (microphone/camera) when putting remote on hold. The stack does not do this as this is where Music on Hold goes, but OpenPhone does not (yet) support that so it needs to shut down the media or the remote non-intuitively can still hear the local user while on hold.

git-svn-id: http://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_10@28579 023b2edf-31b2-4de3-b41e-bca80c47788f