Fix for rejected Invite by Sotel Systems
Sotel Systems challenges an Invite with a 100 Trying and then a 401 Unauthorized.
The 100 Trying has a To: field tag. We save that tag and use it again on
the ensuing Invite. This is wrong and causes Sotel Systems to reject the Invite
with the tag with 481 Call/Transaction Does Not Exist since that tag is no
longer being used. The fix is to not save the tag in the case of a 100 Trying
message.